requirements.txt
flask==3.0.2
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.118.0
boto3==1.34.69
watchtower==3.0.1
python-dotenv==1.0.1
SQLAlchemy==2.0.25
alembic==1.13.1
twilio==8.11.0
gunicorn==21.2.0
black==24.3.0
pylint==3.1.0 